Create HttpFactory::makeRequest()
